CARDS-2742 vocabularies module: component pascalcase, cards-vocabularies.admin-page asset renamed
