+++ modules/vocabularies/src/main/frontend/assets.config
@@ -1,4 +1,4 @@
 // These lines are aggregated as module.exports entries in webpack.config.js file by the webpack_script.py in aggregated-frontend module
 
-['cards-vocabularies.admin-page']: { 'dependOn': ['cards-login.ReLoginDialog'], 'import': './src/vocabulariesAdminPage.jsx' }
-['cards-vocabularies.vocabulariesIcon']: '@mui/icons-material/AccountTree'
+['cards-vocabularies.VocabulariesAdminPage']: { 'dependOn': ['cards-login.ReLoginDialog'], 'import': './src/VocabulariesAdminPage.jsx' }
+['cards-vocabularies.VocabulariesIcon']: '@mui/icons-material/AccountTree'
