+++ modules/vocabularies/src/main/frontend/src/OwlInstaller.jsx
