+++ modules/vocabularies/src/main/frontend/src/VocabularyAction.jsx
